requirements.txt
pillow
pyserial
